requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-jazzmin==3.0.1
idna==3.10
isodate==0.7.2
lxml==5.4.0
mercadopago==2.3.0
pillow==11.2.1
platformdirs==4.3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sqlparse==0.5.3
urllib3==2.4.0
